admin/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.5.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
typer==0.12.5
typing-extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.5.0
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
pika==1.3.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
pytest==8.3.3
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomli==2.0.2
typer==0.12.5
typing-extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
alembic==1.13.3
annotated-types==0.7.0
anyio==4.5.0
bcrypt==4.2.1
certifi==2024.8.30
click==8.1.7
dnspython==2.6.1
email-validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.0
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
iniconfig==2.0.0
jinja2==3.1.4
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
pika==1.3.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-core==2.23.4
pygments==2.18.0
pytest==8.3.3
pytest-env==1.1.5
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
rich==13.9.2
shellingham==1.5.4
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.6
tomli==2.0.2
typer==0.12.5
typing-extensions==4.12.2
uvicorn==0.31.0
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
zipp==3.20.2
